refactor: extract UserProfileTracker
