+++ OptimizelySDK/OptimizelySDK.csproj
@@ -75,6 +75,7 @@
     <Compile Include="AudienceConditions\OrCondition.cs"/>
     <Compile Include="Bucketing\Bucketer.cs"/>
     <Compile Include="Bucketing\Decision.cs"/>
+    <Compile Include="Bucketing\UserProfileTracker.cs" />
     <Compile Include="ClientConfigHandler.cs"/>
     <Compile Include="Entity\Attribute.cs"/>
     <Compile Include="Entity\Audience.cs"/>
